# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HYDROPY SMS Gateway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalDevices: number;
  onlineDevices: number;
  offlineDevices: number;
  smsSentToday: number;
  smsPending: number;
  smsFailed: number;
  activeCampaigns: number;
  totalCampaigns: number;
  queuedMessages: number;
  successRate: number;
}

export type DeviceStatus = typeof DeviceStatus[keyof typeof DeviceStatus];


export const DeviceStatus = {
  online: 'online',
  offline: 'offline',
  disabled: 'disabled',
} as const;

export interface Device {
  id: number;
  name: string;
  deviceId: string;
  status: DeviceStatus;
  /** @nullable */
  battery?: number | null;
  /** @nullable */
  androidVersion?: string | null;
  /** @nullable */
  phoneModel?: string | null;
  /** @nullable */
  simInfo?: string | null;
  /** @nullable */
  signalStrength?: number | null;
  /** @nullable */
  lastHeartbeat?: string | null;
  smsCount?: number;
  /** @nullable */
  currentTask?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface DeviceInput {
  name: string;
  deviceId: string;
  phoneModel?: string;
  androidVersion?: string;
}

export type DeviceUpdateStatus = typeof DeviceUpdateStatus[keyof typeof DeviceUpdateStatus];


export const DeviceUpdateStatus = {
  online: 'online',
  offline: 'offline',
  disabled: 'disabled',
} as const;

export interface DeviceUpdate {
  name?: string;
  status?: DeviceUpdateStatus;
}

export interface TestSmsInput {
  phoneNumber: string;
  message: string;
}

export interface PairInput {
  pairCode: string;
  deviceId: string;
  phoneModel: string;
  androidVersion: string;
  simInfo?: string;
}

export interface PairResult {
  success: boolean;
  deviceId: number;
  token: string;
  wsUrl: string;
}

export interface PairCodeResult {
  pairCode: string;
  expiresAt: string;
}

export interface ApkVersion {
  id: number;
  version: string;
  filename: string;
  /** @nullable */
  size?: number | null;
  isLatest: boolean;
  downloadUrl?: string;
  /** @nullable */
  changelog?: string | null;
  createdAt: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  running: 'running',
  paused: 'paused',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Campaign {
  id: number;
  name: string;
  status: CampaignStatus;
  message: string;
  totalContacts: number;
  sentCount: number;
  failedCount: number;
  pendingCount: number;
  assignedDevices?: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface CampaignInput {
  name: string;
  message: string;
}

export interface CampaignUpdate {
  name?: string;
  message?: string;
}

export type ContactsUploadContactsItem = {
  phoneNumber: string;
  name?: string;
};

export interface ContactsUpload {
  contacts: ContactsUploadContactsItem[];
}

export interface ContactsUploadResult {
  imported: number;
  skipped: number;
}

export type SmsMessageStatus = typeof SmsMessageStatus[keyof typeof SmsMessageStatus];


export const SmsMessageStatus = {
  CREATED: 'CREATED',
  QUEUED: 'QUEUED',
  ASSIGNED: 'ASSIGNED',
  SENT_TO_DEVICE: 'SENT_TO_DEVICE',
  SENDING: 'SENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
} as const;

export interface SmsMessage {
  id: number;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  deviceId?: number | null;
  phoneNumber: string;
  message: string;
  status: SmsMessageStatus;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  resultAt?: string | null;
}

export interface SmsListResponse {
  messages: SmsMessage[];
  total: number;
}

export interface SingleSmsInput {
  phoneNumber: string;
  message: string;
  deviceId?: number;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  device_connected: 'device_connected',
  device_disconnected: 'device_disconnected',
  campaign_started: 'campaign_started',
  campaign_completed: 'campaign_completed',
  sms_assigned: 'sms_assigned',
  sms_sent: 'sms_sent',
  sms_failed: 'sms_failed',
  campaign_paused: 'campaign_paused',
  campaign_resumed: 'campaign_resumed',
  device_error: 'device_error',
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  /** @nullable */
  deviceId?: number | null;
  /** @nullable */
  deviceName?: string | null;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  campaignName?: string | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export type DeviceLogLevel = typeof DeviceLogLevel[keyof typeof DeviceLogLevel];


export const DeviceLogLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface DeviceLog {
  id: number;
  deviceId: number;
  level: DeviceLogLevel;
  message: string;
  /** @nullable */
  data?: string | null;
  createdAt: string;
}

export type SystemLogLevel = typeof SystemLogLevel[keyof typeof SystemLogLevel];


export const SystemLogLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface SystemLog {
  id: number;
  level: SystemLogLevel;
  category: string;
  message: string;
  /** @nullable */
  data?: string | null;
  createdAt: string;
}

export type SystemHealthDatabase = typeof SystemHealthDatabase[keyof typeof SystemHealthDatabase];


export const SystemHealthDatabase = {
  healthy: 'healthy',
  degraded: 'degraded',
  down: 'down',
} as const;

export type SystemHealthBackend = typeof SystemHealthBackend[keyof typeof SystemHealthBackend];


export const SystemHealthBackend = {
  healthy: 'healthy',
  degraded: 'degraded',
  down: 'down',
} as const;

export type SystemHealthWebsocket = typeof SystemHealthWebsocket[keyof typeof SystemHealthWebsocket];


export const SystemHealthWebsocket = {
  healthy: 'healthy',
  degraded: 'degraded',
  down: 'down',
} as const;

export type SystemHealthQueueWorker = typeof SystemHealthQueueWorker[keyof typeof SystemHealthQueueWorker];


export const SystemHealthQueueWorker = {
  healthy: 'healthy',
  degraded: 'degraded',
  down: 'down',
} as const;

export interface SystemHealth {
  database: SystemHealthDatabase;
  backend: SystemHealthBackend;
  websocket: SystemHealthWebsocket;
  queueWorker: SystemHealthQueueWorker;
  connectedDevices: number;
  uptime: number;
}

export interface ActionResult {
  success: boolean;
  message: string;
}

export type ListDevicesParams = {
status?: ListDevicesStatus;
};

export type ListDevicesStatus = typeof ListDevicesStatus[keyof typeof ListDevicesStatus];


export const ListDevicesStatus = {
  online: 'online',
  offline: 'offline',
  all: 'all',
} as const;

export type ListCampaignsParams = {
status?: string;
};

export type ListSmsMessagesParams = {
status?: string;
campaign_id?: number;
limit?: number;
offset?: number;
};

export type GetActivityParams = {
limit?: number;
};

export type GetApiLogsParams = {
limit?: number;
};

export type GetWebSocketLogsParams = {
limit?: number;
};