# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HYDROPY SMS Gateway API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/start:
    post:
      operationId: startCampaign
      tags: [campaigns]
      summary: Start a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a running campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /campaigns/{id}/resume:
    post:
      operationId: resumeCampaign
      tags: [campaigns]
      summary: Resume a paused campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /campaigns/{id}/cancel:
    post:
      operationId: cancelCampaign
      tags: [campaigns]
      summary: Cancel a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /campaigns/{id}/contacts:
    post:
      operationId: uploadContacts
      tags: [campaigns]
      summary: Upload contacts CSV for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactsUpload"
      responses:
        "200":
          description: Contacts uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactsUploadResult"
...
components:
  schemas:
...
    Campaign:
      type: object
      required:
        - id
        - name
        - status
        - message
        - totalContacts
        - sentCount
        - failedCount
        - pendingCount
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [draft, running, paused, completed, cancelled]
        message:
          type: string
        totalContacts:
          type: integer
        sentCount:
          type: integer
        failedCount:
          type: integer
        pendingCount:
          type: integer
        assignedDevices:
          type: integer
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    CampaignInput:
      type: object
      required:
        - name
        - message
      properties:
        name:
          type: string
        message:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
    ContactsUpload:
      type: object
      required:
        - contacts
      properties:
        contacts:
          type: array
          items:
            type: object
            required:
              - phoneNumber
            properties:
              phoneNumber:
                type: string
              name:
                type: string
    ContactsUploadResult:
      type: object
      required:
        - imported
        - skipped
      properties:
        imported:
          type: integer
        skipped:
          type: integer
...
    ActionResult:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string